Move processors to a new file.
